requirements.txt
pystray
pillow
psutil
hid
requests==2.28.2